Revert "reorganize transforms"
This reverts commit f49607f476279d006c9f6c7b29c0e0552560ff1c.
