+++ serotiny/datamodules/dataframe/utils.py
@@ -23,7 +23,7 @@ def get_canonical_split_name(split):
 
 
 def get_dataset(dataframe, transform, split, cache_dir=None):
-    data = _DataframeWrapper(dataframe)
+    data = list(dataframe.to_dict("records"))
     if cache_dir is not None and split in ("train", "val"):
         return PersistentDataset(data, transform=transform, cache_dir=cache_dir)
     return Dataset(data, transform=transform)
@@ -139,14 +139,3 @@ def parse_transforms(transforms):
         transforms["predict"] = transforms["test"]
 
     return transforms
-
-
-class _DataframeWrapper:
-    def __init__(self, df):
-        self.df = df
-
-    def __len__(self):
-        return len(self.df)
-
-    def __getitem__(self, ix):
-        return self.df.iloc[ix].to_dict()
